Add atribute long

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -7,7 +7,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned int tmp;
+	unsigned long int tmp;
 	int i;
 
 	if (n == 0)
